requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via starlette
dnspython==2.6.1
    # via pymongo
fastapi==0.110.2
    # via -r requirements.in
idna==3.7
    # via anyio
motor==3.3.1
    # via -r requirements.in
pydantic==2.7.1
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.18.2
    # via pydantic
pymongo==4.5.0
    # via
    #   -r requirements.in
    #   motor
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.37.2
    # via fastapi
typing-extensions==4.11.0
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic
    #   pydantic-core
